fix: replace numericupdown value type to decimal (responding to avalonia11 changes)

+++ Converters/MathAddConverter.cs
@@ -8,11 +8,11 @@ public class MathAddConverter : IValueConverter
 {
     public object? Convert(object? value, Type targetType, object? parameter, CultureInfo culture)
     {
-        return (double?)value + (double?)parameter;
+        return (decimal?)value + (decimal?)parameter;
     }
 
     public object? ConvertBack(object? value, Type targetType, object? parameter, CultureInfo culture)
     {
-        return (double?)value - (double?)parameter;
+        return (decimal?)value - (decimal?)parameter;
     }
 }
